requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
Django==3.0.5
filetype==1.0.6
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
Pillow==7.1.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
torch==1.4.0
torchvision==0.5.0
typed-ast==1.4.1
wrapt==1.11.2
